Improved alien movement, began shooting system
Most of this commit lies in the new firing system, although there is no
actual "killing" of anything, the framework for that is there and the
feature will likely be present after turrets can shoot and enemies
spawn. There is a graphic representation of the user's firing, all that
is left is the physical raycast to detect a hit.

Turrets now track the nearest enemy.

Aliens have had their movement refined to appear more natural.

A "main" script has been added to aliens, this script manages the List
of aliens which is passed to the turret targeting script. The main
script may be used in the future for other purposes.