Actual Shooting Mechanic Implemented
Firing now detects hits and draws the ray appropriately.

Details: Created a Physics.Raycast using the ray from the camera with a
given range to detect whether there is a hit, if there is a hit, the
line renderer will draw to the hit.point position, otherwise, the full
ray will be drawn as there is nothing in its path.

There is no damage, death, respawn or enemy spawn mechanics and only the
player fires for now.